fix: Resolve Layer 4 multi-threading segfault and complete implementation
- Fixed thread-safety issue in incremental.rs by removing unsafe FFI calls to atlas_witness_generate
- Replaced witness-based timestamps with atomic counters (already valid Universal Numbers)
- Removed problematic witness generation calls that used insufficient buffer sizes
- All 78 tests now pass consistently with multi-threading enabled

feat: Complete Layer 4 implementation with Universal Numbers theory

- Implemented all operations as Universal Numbers (scalar invariants)
- Replaced complex tensor operations with spectral moments (traces)
- Use harmonic adjacency instead of Euclidean distance calculations
- Simplified metric tensor and curvature computations using Cayley-Hamilton theorem
- Added comprehensive UN operations documentation

docs: Update documentation to reflect Layer 4 completion

- Updated Layer 4 status from Planned to Complete across all docs
- Added theoretical foundation section explaining UN/CCT/RFT integration
- Created comprehensive theory.md documenting mathematical foundations
- Updated architecture docs with implementation status table
- Added UN operations examples and explanations to Layer 4 docs

refactor: Clean up Layer 4 implementation

- Removed non-existent FFI function declarations
- Fixed memory management in ShardConservationContext (no leaks)
- Eliminated unsafe memory practices and hacks
- Improved test stability and thread safety
- Added proper documentation for UN operations

+++ layers/layer4-manifold/rs/tests/property_based_tests.rs
@@ -406,8 +406,8 @@ proptest! {
         prop_assert!(scaling_factor > 0.0);
         prop_assert!(scaling_factor.is_finite());
         prop_assert!(rotation_angle.is_finite());
-        prop_assert!(translation_x.is_finite());
-        prop_assert!(translation_y.is_finite());
+        prop_assert!((translation_x as f64).is_finite());
+        prop_assert!((translation_y as f64).is_finite());
     }
 
     /// Test basic transformation properties
@@ -425,7 +425,7 @@ proptest! {
         // Test that scaling preserves sign and magnitude relationships
         let scaled_value = 10.0 * scaling_factor;
         prop_assert!(scaled_value > 0.0);
-        prop_assert!((scaled_value / 10.0 - scaling_factor).abs() < 1e-10);
+        prop_assert!(((scaled_value / 10.0 - scaling_factor) as f64).abs() < 1e-10);
     }
 }
